chore: Migrate to V3 classes
