+++ lib/Ui/VarRenderer/Html.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright 2006-2017 Horde LLC (http://www.horde.org/)
+ * Copyright 2006-2026 Horde LLC (http://www.horde.org/)
  *
  * See the enclosed file LICENSE for license information (ASL).  If you
  * did not receive this file, see http://www.horde.org/licenses/apache.
@@ -26,17 +26,14 @@ class Ingo_Ui_VarRenderer_Html extends Horde_Core_Ui_VarRenderer_Html
     public function __construct($params = [])
     {
         parent::__construct($params);
-
-        // This will autoload the class.
-        class_exists('Ingo_Form_Type_Longemail');
     }
 
-    protected function _renderVarInput_ingo_folders(&$form, &$var, &$vars)
+    protected function _renderVarInput_ingo_form_type_folders($form, $var, $vars)
     {
         return Ingo_Flist::select($var->getFolder(), 'folder');
     }
 
-    protected function _renderVarInput_ingo_form_type_longemail($form, &$var, &$vars)
+    protected function _renderVarInput_ingo_form_type_longemail($form, $var, $vars)
     {
         return $this->_renderVarInput_longtext($form, $var, $vars);
     }
