+++ src/Form/V3/LongemailVariable.php
@@ -1,13 +1,18 @@
 <?php
 
+namespace Ingo\Form\V3;
+
+use Horde\Form\V3\LongtextVariable;
+use Horde_Variables;
+
 /**
- * Copyright 2013-2017 Horde LLC (http://www.horde.org/)
+ * Copyright 2013-2026 Horde LLC (http://www.horde.org/)
  *
  * See the enclosed file LICENSE for license information (ASL).  If you
  * did not receive this file, see http://www.horde.org/licenses/apache.
  *
  * @category  Horde
- * @copyright 2013-2017 Horde LLC
+ * @copyright 2013-2026 Horde LLC
  * @license   http://www.horde.org/licenses/apache ASL
  * @package   Ingo
  */
@@ -18,22 +23,21 @@
  * @author    Michael Slusarz <slusarz@horde.org>
  * @author    yann@pleiades.fr.eu.org
  * @category  Horde
- * @copyright 2013-2017 Horde LLC
+ * @copyright 2013-2026 Horde LLC
  * @license   http://www.horde.org/licenses/apache ASL
  * @package   Ingo
  */
-class Ingo_Form_Type_Longemail extends Horde_Form_Type_longtext
+class LongemailVariable extends LongtextVariable
 {
     /**
      */
-    public function isValid($var, $vars, $value, $message)
+    public function isValid(Horde_Variables $vars, $value): bool
     {
-        $value = trim($value);
+        $value = trim((string)$value);
 
-        if (empty($value)) {
-            if ($var->isRequired()) {
-                $this->message = _("This field is required.");
-                return false;
+        if ($value === '') {
+            if ($this->isRequired()) {
+                return $this->invalid(_("This field is required."));
             }
             return true;
         }
@@ -51,16 +55,15 @@ class Ingo_Form_Type_Longemail extends Horde_Form_Type_longtext
             }
         }
 
-        if (count($invalid)) {
-            $this->message = sprintf(
+        if ($invalid) {
+            return $this->invalid(sprintf(
                 ngettext(
                     _("\"%s\" is not a valid email address."),
                     _("\"%s\" are not valid email addresses."),
                     count($invalid)
                 ),
                 implode(', ', $invalid)
-            );
-            return false;
+            ));
         }
 
         return true;
